set error handling

+++ routes/goals.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const mysql = require('mysql');
+const mysql2 = require('mysql2/promise');
 const { ensureAuthenticated  }= require('../config/auth');
 
 const db = mysql.createConnection({
@@ -9,6 +10,13 @@ const db = mysql.createConnection({
     password : process.env.password,
     database : 'GoalDb'
 });
+const db2 = mysql2.createConnection({
+    host     : 'localhost',
+    user     : 'root',
+    password : process.env.password,
+    database : 'GoalDb'
+});
+
 
 
 // Create New Goal 
@@ -29,7 +37,7 @@ const db = mysql.createConnection({
     let sql = 'INSERT INTO goals SET ?';   // ?means the db.query second param
     let query = db.query(sql, goal, (err, result) => {
         if(err) throw err;
-        let sql2 = `INSERT INTO goalOrderList SET goal_id = ${result.insertId}`
+        let sql2 = `INSERT INTO goalOrderList SET goal_id = ${result.insertId}, user_id = ${req.params.userId};`
         let query2 = db.query(sql2, (err, result) => {});
         res.send(result);
     })
@@ -44,7 +52,7 @@ const db = mysql.createConnection({
  */ 
 router.get('/:userId/goals/',ensureAuthenticated, (req, res) => {
     //
-    let sql = `SELECT * FROM goals JOIN goalOrderList ON goals.goal_id = goalOrderList.goal_id WHERE user_id = ${req.params.userId}`;   
+    let sql = `SELECT * FROM goals JOIN goalOrderList ON goals.goal_id = goalOrderList.goal_id AND goals.user_id = goalOrderList.user_id WHERE goals.user_id = ${req.params.userId}`;   
     let query = db.query(sql, (err, results) => {
         if(err) throw err;
         res.send(results)
@@ -58,45 +66,43 @@ router.get('/:userId/goals/',ensureAuthenticated, (req, res) => {
  * @param {num} ?goalId
  * @return {JSON} goal delete sucess info
  */ 
-router.delete('/:userId/goals/:goalId', ensureAuthenticated, (req, res) => {
-    let sql1 = `SELECT goal_OrderId FROM goalOrderList WHERE goal_id = ${req.params.goalId}`
-    let query1 = db.query(sql1, (err, result) => {
-        if(err) throw err;
-        const deleteItemOrder = result[0].goal_OrderId;
-        let theLastOrderId = 0;
-        
+router.delete('/:userId/goals/:goalId', ensureAuthenticated, async (req, res) => {
+    const con = await db2;
+    try{
+        // Search the goal_OrderId of the deleted goalId
+        let sql1 = `SELECT goal_OrderId FROM goalOrderList WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`
+        let query1 = await con.query(sql1);
+
+        const deleteItemOrder = query1[0][0].goal_OrderId;
+
+        // Delete item form goals table
         let sql2 = `DELETE FROM goals WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}; `;   
-        let query2 = db.query(sql2, (err, result) => {
-            if(err) throw err;
-        })
-        // delete the order item
-        let sql3 = `DELETE FROM goalOrderList WHERE goal_orderId = ${deleteItemOrder} `;
-        let query3 = db.query(sql3, (err, result) => {
-            if(err) throw err;
-        })
+        let query2 = await con.query(sql2);
+
+        // Delete item from goalOrderList table 
+        let sql3 = `DELETE FROM goalOrderList WHERE goal_orderId = ${deleteItemOrder} AND user_id = ${req.params.userId};`;
+        let query3 = await con.query(sql3);
 
         // Rearrange new order
         // 1 count 
-        let sql4 = `SELECT COUNT(goal_orderId) FROM goalOrderList;`;
-        let query4 = db.query(sql4,  (err, result) => {
-            if(err) throw err;
-            theLastOrderId = result[0]['COUNT(goal_orderId)'];
+        let theLastOrderId = 0;
+        let sql4 = `SELECT COUNT(goal_orderId) FROM goalOrderList; AND user_id = ${req.params.userId};`;
+        let query4 = await con.query(sql4)
 
-            for (let i = deleteItemOrder + 1; i <= theLastOrderId + 1; i++){
-                // Find the Id of the goal_ording
-                let sql5 = `SELECT * FROM goalOrderList WHERE goal_orderId = ${i}`;
-                db.query(sql5, (err, result) => {
-                    let currentGoalId = result[0].goal_id ;
-                    let sql6 = `UPDATE goalOrderList SET goal_orderId = ${i - 1} WHERE goal_id = ${currentGoalId}`
-                    db.query(sql6, (err, result) => {
-                        console.log(currentGoalId);
-                   })
-                })
-            }
-        })
+        theLastOrderId = query4[0][0]['COUNT(goal_orderId)'];
 
-        res.send('hihi')
-    })
+        for (let i = deleteItemOrder + 1; i <= theLastOrderId + 1; i++){
+            
+            // Find the Id of the goal_ording
+            let sql5 = `SELECT * FROM goalOrderList WHERE goal_orderId = ${i} AND user_id = ${req.params.userId};`;
+            let query5 = await con.query(sql5)
+            let currentGoalId = query5[0][0].goal_id ;
+            let sql6 = `UPDATE goalOrderList SET goal_orderId = ${i - 1} WHERE goal_id = ${currentGoalId} AND user_id = ${req.params.userId};`
+            await con.query(sql6)
+        }     
+        res.send([{msg: 'Delete Success'}]);
+    }    
+    catch(e){ res.send(e)};
 })
 
 // Update the goal
@@ -108,16 +114,39 @@ router.delete('/:userId/goals/:goalId', ensureAuthenticated, (req, res) => {
  * @param {string} goalDescription
  * @param {boolean} goalCompleted
  * @return {JSON} goal update sucess info
+ * not block not exsit goadId
  */ 
-router.put('/:userId/goals/:goalId', ensureAuthenticated, (req, res) => {
-    let sql = `UPDATE goals SET goal_name = '${req.body.goalName}', goal_description = '${req.body.goalDescription}',goal_completed = ${req.body.goalCompleted}  WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`;   
-    let query = db.query(sql, (err, result) => {
-        if(err) throw err;
-        console.log(result);
-        res.send(result)
-    })
+router.put('/:userId/goals/:goalId', ensureAuthenticated, async (req, res) => {
+    try{
+        const con = await db2;
+        let sql = `UPDATE goals SET goal_name = '${req.body.goalName}', goal_description = '${req.body.goalDescription}',goal_completed = ${req.body.goalCompleted}  WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`;   
+        let query = await con.query(sql);
+        res.send([{msg : "Upate Success"}])
+    }
+    catch(e){res.send(e)}
 })
 
+
+// Update the goal order
+/**
+ * @METH PUT
+ * @param {num} ?userId
+ * @param {num} ?goalId
+ * @param {num} GoadFrom
+ * @param {num} GoadTo
+ * @return {JSON} goal update sucess info
+ * First index = 1
+ * not block not exsit goadId
+ */ 
+ router.put('/:userId/goals/:goalId/updateOrder', ensureAuthenticated, async (req, res) => {
+    try{
+        const con = await db2;
+        let sql = `UPDATE goalOrderList SET goal_orderId = ${req.body.GoadTo}  WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`;   
+        let query = await con.query(sql);
+        res.send([{msg : "Upate Success"}])
+    }
+    catch(e){res.send(e)}
+})
 
 // Get goal detail, subGoal
 /**
@@ -126,12 +155,14 @@ router.put('/:userId/goals/:goalId', ensureAuthenticated, (req, res) => {
  * @param {num} ?goalId
  * @return {array} subgoals
  */ 
-router.get('/:userId/goals/:goalId', ensureAuthenticated,(req, res) => {
-    let sql = `SELECT * FROM subgoals WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`;   
-    let query = db.query(sql, (err, results) => {
-        if(err) throw err;        
-        res.send(results)
-    })
+router.get('/:userId/goals/:goalId', ensureAuthenticated,async (req, res) => {
+    try{
+        const con = await db2;
+        let sql = `SELECT * FROM subgoals WHERE goal_id = ${req.params.goalId} AND user_id = ${req.params.userId}`;   
+        let query = await con.query(sql)       
+        res.send(query[0])
+    }
+    catch(e){res.send(e)}
 })
 
 // Creat new subgoal
@@ -144,22 +175,22 @@ router.get('/:userId/goals/:goalId', ensureAuthenticated,(req, res) => {
  * @param {int} subGoalPredictTime
  * @return {JSON} subgoal creat sucess info
  */ 
- router.post('/:userId/goals/:goalId', ensureAuthenticated,(req, res) => {
-    let subgoal = {
-        user_id: req.params.userId,
-        goal_id: req.params.goalId,
-        subgoal_name: req.body.subGoalName, 
-        subgoal_description: req.body.subGoalDescription,
-        subgoal_predict_time: req.body.subGoalPredictTime,
-        subgoal_completed: false
-    };
-    console.log(req.body);
-    let sql = 'INSERT INTO subgoals SET ?';   // ?means the db.query second param
-    let query = db.query(sql, subgoal, (err, result) => {
-        if(err) throw err;
-        console.log(result.insertId);
-        res.send(result);
-    })
+ router.post('/:userId/goals/:goalId', ensureAuthenticated, async(req, res) => {
+    try{
+        const con = await db2;
+        let subgoal = {
+            user_id: req.params.userId,
+            goal_id: req.params.goalId,
+            subgoal_name: req.body.subGoalName, 
+            subgoal_description: req.body.subGoalDescription,
+            subgoal_predict_time: req.body.subGoalPredictTime,
+            subgoal_completed: false
+        };
+        let sql = 'INSERT INTO subgoals SET ?';   // ?means the db.query second param
+        let query = await con.query(sql, subgoal)
+        res.send(query[0]);    
+    }
+    catch(e){res.send(e)}
 })
